[specifications] make all specs work with `$ fontbakery check-spec`
