+++ Lib/fontbakery/specifications/gpos.py
@@ -5,6 +5,10 @@ from fontbakery.callable import check, condition
 from fontbakery.checkrunner import FAIL, PASS, WARN
 from fontbakery.message import Message
 
+from .shared_conditions import ligatures
+from .googlefonts_shared_conditions import whitelist_librebarcode
+# flake8 F401, F811:
+(ligatures, whitelist_librebarcode)
 
 @condition
 def has_kerning_info(ttFont):
