+++ Lib/fontbakery/specifications/name.py
@@ -7,6 +7,11 @@ from fontbakery.constants import CRITICAL
 from fontbakery.message import Message
 
 
+from .shared_conditions import seems_monospaced, monospace_stats
+from .googlefonts_shared_conditions import whitelist_librebarcode
+# flake8 F401, F811:
+(seems_monospaced, monospace_stats, whitelist_librebarcode)
+
 @check(id='com.google.fonts/check/031', priority=CRITICAL)
 def com_google_fonts_check_031(ttFont):
   """Description strings in the name table must not contain copyright info."""
